apply isOptionUnique callback to GroupsElement
